requirements.txt
requests>=2.31.0
urllib3>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
dateparser>=1.2.0
thefuzz>=0.19.0
python-Levenshtein>=0.21.0
regex>=2024.9.11
groq>=0.4.0
google-play-scraper>=1.2.0
sqlalchemy>=2.0.0
alembic>=1.13.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
email-validator>=2.0.0
apscheduler>=3.10.0
pytz>=2023.3

